backend/requirements/base.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers
djangorestframework-simplejwt
Pillow
django-filter
django-extensions
